# shared/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Dependency Upgrade Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RepositoryUploadInput {
  file: string;
}

export interface GithubRepositoryInput {
  url: string;
}

export type RepositorySource = typeof RepositorySource[keyof typeof RepositorySource];


export const RepositorySource = {
  zip: 'zip',
  github: 'github',
} as const;

export type RepositoryPackageManager = typeof RepositoryPackageManager[keyof typeof RepositoryPackageManager];


export const RepositoryPackageManager = {
  npm: 'npm',
  pnpm: 'pnpm',
  yarn: 'yarn',
  unsupported: 'unsupported',
} as const;

export type RepositoryStatus = typeof RepositoryStatus[keyof typeof RepositoryStatus];


export const RepositoryStatus = {
  analyzed: 'analyzed',
  invalid: 'invalid',
} as const;

export type DependencySection = typeof DependencySection[keyof typeof DependencySection];


export const DependencySection = {
  dependencies: 'dependencies',
  devDependencies: 'devDependencies',
  peerDependencies: 'peerDependencies',
  optionalDependencies: 'optionalDependencies',
} as const;

export interface Dependency {
  name: string;
  version: string;
  section: DependencySection;
}

export interface Repository {
  id: string;
  name: string;
  source: RepositorySource;
  language: string;
  packageManager: RepositoryPackageManager;
  hasPackageJson: boolean;
  /** @nullable */
  lockfile: string | null;
  /** @nullable */
  framework: string | null;
  dependencies: Dependency[];
  scripts: string[];
  status: RepositoryStatus;
  createdAt: string;
}

export type MigrationInputMode = typeof MigrationInputMode[keyof typeof MigrationInputMode];


export const MigrationInputMode = {
  agentic: 'agentic',
  baseline: 'baseline',
} as const;

export interface MigrationInput {
  repositoryId: string;
  /** @minLength 1 */
  dependency: string;
  /** @pattern ^[0-9]+$ */
  targetMajor: string;
  mode?: MigrationInputMode;
}

export type MigrationMode = typeof MigrationMode[keyof typeof MigrationMode];


export const MigrationMode = {
  agentic: 'agentic',
  baseline: 'baseline',
} as const;

export type MigrationStatus = typeof MigrationStatus[keyof typeof MigrationStatus];


export const MigrationStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type MigrationTests = typeof MigrationTests[keyof typeof MigrationTests];


export const MigrationTests = {
  pass: 'pass',
  fail: 'fail',
  skipped: 'skipped',
  running: 'running',
} as const;

export type MigrationBuild = typeof MigrationBuild[keyof typeof MigrationBuild];


export const MigrationBuild = {
  pass: 'pass',
  fail: 'fail',
  skipped: 'skipped',
  running: 'running',
} as const;

export type MigrationTypecheck = typeof MigrationTypecheck[keyof typeof MigrationTypecheck];


export const MigrationTypecheck = {
  pass: 'pass',
  fail: 'fail',
  skipped: 'skipped',
  running: 'running',
} as const;

export type MigrationLint = typeof MigrationLint[keyof typeof MigrationLint];


export const MigrationLint = {
  pass: 'pass',
  fail: 'fail',
  skipped: 'skipped',
  running: 'running',
} as const;

export type MigrationAgentStateFileChangesItem = {
  path?: string;
  action?: string;
};

export interface AgentToolCall {
  timestamp: string;
  tool: string;
  inputSummary: string;
  resultSummary: string;
  success: boolean;
  durationMs: number;
  /** @nullable */
  errorType?: string | null;
}

/**
 * Public agent activity recorded by the coding agent.
 * @nullable
 */
export type MigrationAgentState = {
  status?: string;
  currentAction?: string;
  toolCalls?: AgentToolCall[];
  filesInspected?: string[];
  filesModified?: string[];
  patchesApplied?: number;
  /** @nullable */
  planSummary?: string | null;
  fileChanges?: MigrationAgentStateFileChangesItem[];
  /** @nullable */
  agentSummary?: string | null;
  /** @nullable */
  error?: string | null;
} | null;

export interface Migration {
  id: string;
  repositoryId: string;
  repositoryName: string;
  dependency: string;
  oldVersion: string;
  targetVersion: string;
  mode: MigrationMode;
  status: MigrationStatus;
  currentStage: string;
  attemptNumber: number;
  affectedFiles: number;
  affectedUsages: number;
  tests: MigrationTests;
  build: MigrationBuild;
  typecheck: MigrationTypecheck;
  lint: MigrationLint;
  /** @nullable */
  errorCode: string | null;
  createdAt: string;
  updatedAt: string;
  /**
     * Public agent activity recorded by the coding agent.
     * @nullable
     */
  agentState?: MigrationAgentState;
}

export type MigrationEventLevel = typeof MigrationEventLevel[keyof typeof MigrationEventLevel];


export const MigrationEventLevel = {
  info: 'info',
  success: 'success',
  warning: 'warning',
  error: 'error',
} as const;

export interface MigrationEvent {
  id: string;
  migrationId: string;
  timestamp: string;
  level: MigrationEventLevel;
  message: string;
}

export type DiffFileStatus = typeof DiffFileStatus[keyof typeof DiffFileStatus];


export const DiffFileStatus = {
  modified: 'modified',
  added: 'added',
  deleted: 'deleted',
} as const;

export interface DiffFile {
  path: string;
  status: DiffFileStatus;
  patch: string;
  additions: number;
  deletions: number;
}

export interface MigrationDiff {
  filesChanged: number;
  additions: number;
  deletions: number;
  files: DiffFile[];
}

export interface RepositorySummary {
  name: string;
  language: string;
  packageManager: string;
  /** @nullable */
  framework: string | null;
}

export interface ImpactSummary {
  affectedFiles: number;
  affectedUsages: number;
  files: string[];
}

export interface ResearchSource {
  title: string;
  url: string;
  finding: string;
}

export interface Attempt {
  number: number;
  result: string;
  /** @nullable */
  diagnosis: string | null;
  filesChanged: number;
}

export interface MigrationReport {
  migrationId: string;
  status: string;
  summary: string;
  repository: RepositorySummary;
  impact: ImpactSummary;
  sources: ResearchSource[];
  changes: string[];
  attempts: Attempt[];
  remainingIssues: string[];
}

export interface DashboardSummary {
  totalMigrations: number;
  completedMigrations: number;
  runningMigrations: number;
  failedMigrations: number;
  recent: Migration[];
  capabilities: string[];
}