# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dependency Upgrade Agent API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: repositories
    description: Repository import and analysis
  - name: migrations
    description: Dependency migration jobs
  - name: dashboard
    description: Dashboard summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /repositories/upload:
    post:
      operationId: uploadRepository
      tags: [repositories]
      summary: Upload and analyze a ZIP repository
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RepositoryUploadInput"
      responses:
        "201":
          description: Analyzed repository
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "400":
          description: Invalid repository
  /repositories/github:
    post:
      operationId: importGithubRepository
      tags: [repositories]
      summary: Import and analyze a GitHub repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubRepositoryInput"
      responses:
        "201":
          description: Analyzed repository
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "400":
          description: Invalid repository
  /repositories/{id}:
    get:
      operationId: getRepository
      tags: [repositories]
      summary: Get repository analysis
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Repository analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "404":
          description: Repository not found
  /migrations:
    get:
      operationId: listMigrations
      tags: [migrations]
      summary: List recent migrations
      responses:
        "200":
          description: Recent migrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Migration"
    post:
      operationId: createMigration
      tags: [migrations]
      summary: Create and start a migration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MigrationInput"
      responses:
        "202":
          description: Migration started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Migration"
        "400":
          description: Invalid migration request
  /migrations/{id}:
    get:
      operationId: getMigration
      tags: [migrations]
      summary: Get migration job state
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Migration state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Migration"
        "404":
          description: Migration not found
  /migrations/{id}/events:
    get:
      operationId: getMigrationEvents
      tags: [migrations]
      summary: Get migration event log
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Event log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MigrationEvent"
  /migrations/{id}/diff:
    get:
      operationId: getMigrationDiff
      tags: [migrations]
      summary: Get actual migration diff
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Diff grouped by file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationDiff"
  /migrations/{id}/report:
    get:
      operationId: getMigrationReport
      tags: [migrations]
      summary: Get final migration report
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Final report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationReport"
  /migrations/{id}/approve:
    post:
      operationId: approveMigration
      tags: [migrations]
      summary: Approve completed migration
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Approved migration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Migration"
  /migrations/{id}/cancel:
    post:
      operationId: cancelMigration
      tags: [migrations]
      summary: Cancel a running migration
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Cancelled migration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Migration"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    RepositoryUploadInput:
      type: object
      properties:
        file:
          type: string
      required: [file]
    GithubRepositoryInput:
      type: object
      properties:
        url:
          type: string
      required: [url]
    Repository:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        source: { type: string, enum: [zip, github] }
        language: { type: string }
        packageManager: { type: string, enum: [npm, pnpm, yarn, unsupported] }
        hasPackageJson: { type: boolean }
        lockfile: { type: ["string", "null"] }
        framework: { type: ["string", "null"] }
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/Dependency"
        scripts:
          type: array
          items: { type: string }
        status: { type: string, enum: [analyzed, invalid] }
        createdAt: { type: string, format: date-time }
      required: [id, name, source, language, packageManager, hasPackageJson, lockfile, framework, dependencies, scripts, status, createdAt]
    Dependency:
      type: object
      properties:
        name: { type: string }
        version: { type: string }
        section: { type: string, enum: [dependencies, devDependencies, peerDependencies, optionalDependencies] }
      required: [name, version, section]
    MigrationInput:
      type: object
      properties:
        repositoryId: { type: string }
        dependency: { type: string, minLength: 1 }
        targetMajor: { type: string, pattern: "^[0-9]+$" }
        mode: { type: string, enum: [agentic, baseline], default: agentic }
      required: [repositoryId, dependency, targetMajor]
    Migration:
      type: object
      properties:
        id: { type: string }
        repositoryId: { type: string }
        repositoryName: { type: string }
        dependency: { type: string }
        oldVersion: { type: string }
        targetVersion: { type: string }
        mode: { type: string, enum: [agentic, baseline] }
        status: { type: string, enum: [queued, running, completed, failed, cancelled, approved, rejected] }
        currentStage: { type: string }
        attemptNumber: { type: number }
        affectedFiles: { type: number }
        affectedUsages: { type: number }
        tests: { type: string, enum: [pass, fail, skipped, running] }
        build: { type: string, enum: [pass, fail, skipped, running] }
        typecheck: { type: string, enum: [pass, fail, skipped, running] }
        lint: { type: string, enum: [pass, fail, skipped, running] }
        errorCode: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        agentState:
          type: ["object", "null"]
          description: Public agent activity recorded by the coding agent.
          properties:
            status: { type: string }
            currentAction: { type: string }
            toolCalls:
              type: array
              items:
                $ref: "#/components/schemas/AgentToolCall"
            filesInspected:
              type: array
              items: { type: string }
            filesModified:
              type: array
              items: { type: string }
            patchesApplied: { type: number }
            planSummary: { type: ["string", "null"] }
            fileChanges:
              type: array
              items:
                type: object
                properties:
                  path: { type: string }
                  action: { type: string }
            agentSummary: { type: ["string", "null"] }
            error: { type: ["string", "null"] }
      required: [id, repositoryId, repositoryName, dependency, oldVersion, targetVersion, mode, status, currentStage, attemptNumber, affectedFiles, affectedUsages, tests, build, typecheck, lint, errorCode, createdAt, updatedAt]
    AgentToolCall:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        tool: { type: string }
        inputSummary: { type: string }
        resultSummary: { type: string }
        success: { type: boolean }
        durationMs: { type: number }
        errorType: { type: ["string", "null"] }
      required: [timestamp, tool, inputSummary, resultSummary, success, durationMs]
    MigrationEvent:
      type: object
      properties:
        id: { type: string }
        migrationId: { type: string }
        timestamp: { type: string, format: date-time }
        level: { type: string, enum: [info, success, warning, error] }
        message: { type: string }
      required: [id, migrationId, timestamp, level, message]
    MigrationDiff:
      type: object
      properties:
        filesChanged: { type: number }
        additions: { type: number }
        deletions: { type: number }
        files:
          type: array
          items:
            $ref: "#/components/schemas/DiffFile"
      required: [filesChanged, additions, deletions, files]
    DiffFile:
      type: object
      properties:
        path: { type: string }
        status: { type: string, enum: [modified, added, deleted] }
        patch: { type: string }
        additions: { type: number }
        deletions: { type: number }
      required: [path, status, patch, additions, deletions]
    MigrationReport:
      type: object
      properties:
        migrationId: { type: string }
        status: { type: string }
        summary: { type: string }
        repository: { $ref: "#/components/schemas/RepositorySummary" }
        impact: { $ref: "#/components/schemas/ImpactSummary" }
        sources:
          type: array
          items:
            $ref: "#/components/schemas/ResearchSource"
        changes:
          type: array
          items: { type: string }
        attempts:
          type: array
          items:
            $ref: "#/components/schemas/Attempt"
        remainingIssues:
          type: array
          items: { type: string }
      required: [migrationId, status, summary, repository, impact, sources, changes, attempts, remainingIssues]
    RepositorySummary:
      type: object
      properties:
        name: { type: string }
        language: { type: string }
        packageManager: { type: string }
        framework: { type: ["string", "null"] }
      required: [name, language, packageManager, framework]
    ImpactSummary:
      type: object
      properties:
        affectedFiles: { type: number }
        affectedUsages: { type: number }
        files:
          type: array
          items: { type: string }
      required: [affectedFiles, affectedUsages, files]
    ResearchSource:
      type: object
      properties:
        title: { type: string }
        url: { type: string }
        finding: { type: string }
      required: [title, url, finding]
    Attempt:
      type: object
      properties:
        number: { type: number }
        result: { type: string }
        diagnosis: { type: ["string", "null"] }
        filesChanged: { type: number }
      required: [number, result, diagnosis, filesChanged]
    DashboardSummary:
      type: object
      properties:
        totalMigrations: { type: number }
        completedMigrations: { type: number }
        runningMigrations: { type: number }
        failedMigrations: { type: number }
        recent:
          type: array
          items:
            $ref: "#/components/schemas/Migration"
        capabilities:
          type: array
          items: { type: string }
      required: [totalMigrations, completedMigrations, runningMigrations, failedMigrations, recent, capabilities]