# shared/api-zod/src/generated/types/agentToolCall.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Dependency Upgrade Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentToolCall {
  timestamp: Date;
  tool: string;
  inputSummary: string;
  resultSummary: string;
  success: boolean;
  durationMs: number;
  /** @nullable */
  errorType?: string | null;
}