# shared/api-zod/src/generated/types/dependency.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Dependency Upgrade Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { DependencySection } from './dependencySection';

export interface Dependency {
  name: string;
  version: string;
  section: DependencySection;
}