# shared/api-zod/src/generated/types/impactSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Dependency Upgrade Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ImpactSummary {
  affectedFiles: number;
  affectedUsages: number;
  files: string[];
}