# shared/api-zod/src/generated/types/migrationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Dependency Upgrade Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { MigrationInputMode } from './migrationInputMode';

export interface MigrationInput {
  repositoryId: string;
  /** @minLength 1 */
  dependency: string;
  /** @pattern ^[0-9]+$ */
  targetMajor: string;
  mode?: MigrationInputMode;
}