# shared/api-zod/src/generated/types/researchSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Dependency Upgrade Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchSource {
  title: string;
  url: string;
  finding: string;
}